refactor(tui-kit): group component modules
